Move svg generator into separate location
